requirements.txt
bert_score==0.3.13
deepspeed==0.15.1
fast_pytorch_kmeans==0.2.0.1
ffmpeg_python==0.2.0
flash_attn==2.4.2
jsonlines==4.0.0
matplotlib==3.8.3
moverscore==1.0.3
nltk==3.8.1
numerize==0.12
numpy==1.26.4
openai==0.28.0
peft==0.7.1
Pillow==10.3.0
pycocoevalcap==1.2
supar==1.1.4
tqdm==4.65.0
transformers==4.37.2
sentencepiece==0.2.0
av==13.1.0